refactor: move SBOM mock data out of production into Storybook helpers
